+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+## next
+
 ## 1.13.0
 
 - Add support for CodeLLDB extension